Added sell token functionality
