+++ packages/gnosis/customs/omen_tools/component.yaml
@@ -1,15 +1,15 @@
-name: omen_buy_yes
+name: omen_tools
 author: gnosis
 version: 0.1.0
 type: custom
-description: A tool to prepare a request for buying tokens corresponding to outcome Yes on Omen.
+description: Collection of tools to prepare requests for interacting with prediction markets on Omen.
 license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
   __init__.py: bafybeibbn67pnrrm4qm3n3kbelvbs3v7fjlrjniywmw2vbizarippidtvi
   prediction_sum_url_content.py: bafybeieywowx265yycgf5735bw4zyabfy6ivwnntl6smxa2hicktipgeby
 fingerprint_ignore_patterns: []
-entry_point: omen_buy_yes.py
+entry_point: omen_buy_sell.py
 callable: run
 dependencies:
   pydantic:
